# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Neo4j topology graph — nodes, relationships, anomalies, root cause analysis, and blast radius impact.",
        "title": "graphserv API",
        "contact": {
            "name": "Rajesh Kurup",
            "email": "rajeshkurup@live.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/analysis/impact": {
            "get": {
                "description": "Returns reverse-dependency (blast radius) metrics for a node — how many other nodes depend on it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analysis"
                ],
                "summary": "Blast radius / impact analysis",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node label",
                        "name": "label",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node business id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "false",
                        "description": "Use precomputed DEPENDS_ON_TRANSITIVE edges (true/false)",
                        "name": "useTransitive",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/graphstore.ImpactMetrics"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/analysis/root-cause": {
            "post": {
                "description": "Traverses downstream topology edges from a start node to find active anomalies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analysis"
                ],
                "summary": "Root cause analysis",
                "parameters": [
                    {
                        "description": "Start node and traversal parameters",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.rootCauseReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/graphstore.RootCauseResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/anomalies": {
            "post": {
                "description": "Creates or updates an anomaly node and links it to topology nodes via DETECTED_ON relationships",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Anomalies"
                ],
                "summary": "Upsert anomaly with DETECTED_ON edges",
                "parameters": [
                    {
                        "description": "Anomaly properties, target nodes, and relationship properties",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.anomalyReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes": {
            "post": {
                "description": "Creates or merges a node with the given label and properties. Properties must include an \"id\" field.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Ingest (merge) a node",
                "parameters": [
                    {
                        "description": "Node label and properties",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.ingestNodeReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{label}": {
            "get": {
                "description": "Returns all nodes matching the given label with optional limit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "List nodes by label",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node label (e.g. Application, Storage, Network, IncidentTicket, ChangeTicket, RCATicket, Action, Anomaly, Call)",
                        "name": "label",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "Max results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/nodes/{label}/{id}": {
            "get": {
                "description": "Returns a node by its label and business id",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Get a single node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node label",
                        "name": "label",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node business id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Detach-deletes the node and all its relationships",
                "tags": [
                    "Nodes"
                ],
                "summary": "Delete a node",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node label",
                        "name": "label",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node business id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Node deleted"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Merges the given JSON properties onto the node (id cannot be changed)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nodes"
                ],
                "summary": "Update node properties",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Node label",
                        "name": "label",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Node business id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Properties to merge",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.anomalyReq": {
            "type": "object",
            "properties": {
                "anomaly": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "detectedOn": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/graphstore.Endpoint"
                    }
                },
                "relationshipProperties": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "bad request"
                }
            }
        },
        "api.ingestNodeReq": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string",
                    "example": "Application"
                },
                "properties": {
                    "type": "object",
                    "additionalProperties": {}
                }
            }
        },
...
        "api.rootCauseReq": {
            "type": "object",
            "properties": {
                "anomalyStatus": {
                    "type": "string",
                    "example": "active"
                },
                "limit": {
                    "type": "integer",
                    "example": 50
                },
                "maxDepth": {
                    "type": "integer",
                    "example": 5
                },
                "startId": {
                    "type": "string",
                    "example": "app-123"
                },
                "startLabel": {
                    "type": "string",
                    "example": "Application"
                }
            }
        },
        "graphstore.Endpoint": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            }
        },
        "graphstore.ImpactMetrics": {
            "type": "object",
            "properties": {
                "byLabel": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "byRelationshipType": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "dependentCount": {
                    "type": "integer"
                },
                "dependents": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "start": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "usedTransitive": {
                    "type": "boolean"
                }
            }
        },
        "graphstore.RootCauseResult": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "integer"
                },
                "origins": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                }
            }
        }
    }
}